requirements.txt
appdirs==1.4.3
asgiref==3.5.2
bcrypt==3.2.2
cffi==1.15.1
cryptography==37.0.4
dj-database-url==1.0.0
Django==4.0.5
django-admin-honeypot==1.1.0
django-anymail==8.6
django-ckeditor==6.4.2
django-environ==0.9.0
django-etc==1.3.1
django-heroku==0.3.1
django-hitcount==1.3.5
django-js-asset==2.0.0
django-mailjet==0.3.1
django-social-share==2.3.0
django-taggit==3.0.0
gunicorn==20.1.0
mailjet-rest==1.3.4
Pillow==9.2.0
psycopg2==2.9.3
pyparsing==2.2.0
python-decouple==3.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
whitenoise==6.2.0
zxcvbn==4.4.28
